scout/pyproject.toml
[tool.poetry]
name = "paperscout"
version = "0.1.0"
description = "AI-powered paper search and summarization tool"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
langchain-core = "^0.1.0"
aiohttp = "^3.9.0"
python-dotenv = "^1.0.0"
arxiv = "^2.0.0"
PyPDF2 = "^3.0.0"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 